+++ model/main.go
@@ -319,6 +319,9 @@ func migrateDB() error {
 	if err != nil {
 		return err
 	}
+	if err := releaseDeletedUserOAuthBindings(DB); err != nil {
+		return err
+	}
 	refreshTokenModelMappingColumnCache()
 	if err := MigrateDirectCryptoPaymentIndexes(DB); err != nil {
 		return err
@@ -405,6 +408,9 @@ func migrateDBFast() error {
 		if err != nil {
 			return err
 		}
+	}
+	if err := releaseDeletedUserOAuthBindings(DB); err != nil {
+		return err
 	}
 	refreshTokenModelMappingColumnCache()
 	if err := MigrateDirectCryptoPaymentIndexes(DB); err != nil {